deleted old and unused files, some were just renamed
